ipq806x: initialize UART even when console is not enabled
The ipq806x UART is based on the same universal serial port which can
be also configured as i2c or SPI. Configuring it is not a trivial
task, so in case the kernel wants to use earlyprintk() the port needs
to be configured by the firmware.

Invoking uart_init() when the console is not enabled causes include
file collisions, which would require changes to more than 100 files.
Leaving this to another day, rearranging the ipq806x driver to be able
to invoke UART initialization function even when serial console is not
configure in.

Also adding a check to avoid initialization if UART has been already
set up.

BRANCH=storm
BUG=chrome-os-partner:35364
TEST=verified that storm console is still fully operational when
     enabled, and that the kernel boots fine to the serial console
     login prompt even if the firmware console is disabled.

Change-Id: I421acba3edf398d960b5058f15d1abb80ebc7660
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/240516
Reviewed-by: David Hendricks <dhendrix@chromium.org>