broadwell: Turn off panel backlight in S5 SMI handler
In order for some panels to meet spec when the system is put
into S5 by way of power button during firmware (i.e. not by
the OS) then it needs to turn off the backlight and give it
time to turn off before going into S5.

If the OS properly sequences the panel down then the backlight
enable bit will not be set in this step and nothing will happen.

BUG=chrome-os-partner:33994
BRANCH=broadwell
TEST=build and boot on samus

Change-Id: I43c5aee8e32768fc9e82790c9f7ceda0ed17ed13
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/240852
Reviewed-by: Shawn N <shawnn@chromium.org>