samus: Set current backlight PWM value
With recent changes in the 3.14 kernel and the switch to not
using X the panel backlight is not geting turned on until
chrome is started which means the splash screen is not visible.

If we set the backlight PWM in coreboot then it will at least
turn on for the early boot process.

BUG=chrome-os-partner:31549
BRANCH=samus
TEST=boot on samus in normal mode and see the boot splash logo

Change-Id: I622bef8af9bb6b753fe228b33ecdc4aae76af131
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/240853
Reviewed-by: Shawn N <shawnn@chromium.org>