ACPI: Get code to match header file
The build breaks when CONFIG_HAVE_ACPI_RESUME is not selected because
the arch/acpi.h file only defines the prototype for
mainboard_suspend_resume when CONFIG_HAVE_ACPI_RESUME is selected.  Fix
the code not define the mainboard_suspend_resume routine when
CONFIG_HAVE_ACPI_RESUME is not selected.

BRANCH=none
BUG=None
TEST=Build with and without CONFIG_HAVE_ACPI_RESUME

Change-Id: I126e1116114a6b92c8a42d0cf5fce076cb6f26a4
Signed-off-by: Lee Leahy <leroy.p.leahy@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/241071
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>