libpayload: sync arch/arm/cache.c with coreboot
There was a recent patch by Deepa Dinamani applied to coreboot's
cache.c which fixed a bug that occurred when icache is on but dcache
is off ("arch: armv7: Fix cache sync instructions."). Although this
bug is not likely to be encountered by the time libpayload is run,
it's worth applying it to keep things in sync.

BUG=none
BRANCH=none
TEST=n/a since we have icache and dcache enabled on all ARM platforms
when libpayload is run.

Signed-off-by: David Hendricks <dhendrix@chromium.org>
Change-Id: I4ab0d97ef3a97dcd0fa96e10273c3b32486e0b40
Reviewed-on: https://chromium-review.googlesource.com/243276
Reviewed-by: Julius Werner <jwerner@chromium.org>